record_id should be a bigint

+++ spark/common/medicalclaims_common_model.sql
@@ -1,6 +1,6 @@
 DROP TABLE IF EXISTS {table_name};
 CREATE TABLE {table_name} (
-        record_id                          int,
+        record_id                          bigint,
         claim_id                           string,
         hvid                               string,
         created                            date,
